+++ src/plasma/plasma_manager.cc
@@ -238,7 +238,7 @@ struct PlasmaManagerState {
   AvailableObject *local_available_objects;
   /** The time (in milliseconds since the Unix epoch) when the most recent
    *  heartbeat was sent. */
-  int previous_heartbeat_time;
+  int64_t previous_heartbeat_time;
 };
 
 PlasmaManagerState *g_manager_state = NULL;
@@ -1597,11 +1597,11 @@ int heartbeat_handler(event_loop *loop, timer_id id, void *context) {
   PlasmaManagerState *state = (PlasmaManagerState *) context;
 
   /* Check that the last heartbeat was not sent too long ago. */
-  int current_time = current_time_ms();
+  int64_t current_time = current_time_ms();
   CHECK(current_time >= state->previous_heartbeat_time);
   if (current_time - state->previous_heartbeat_time >
       NUM_HEARTBEATS_TIMEOUT * HEARTBEAT_TIMEOUT_MILLISECONDS) {
-    LOG_FATAL("The last heartbeat was sent %d milliseconds ago.",
+    LOG_FATAL("The last heartbeat was sent %" PRId64 " milliseconds ago.",
               current_time - state->previous_heartbeat_time);
   }
   state->previous_heartbeat_time = current_time;
